Add major update to population, GDP distribution project
Add a helpful README to provide overview of the article, data source
and notes on the changes made to the visualizations.

Add SVGs for the visualizations. Add the resized and cropped versions
of the visualizations in PNGs.

Add visualizations for the GRDP per capita and incoming lifetime
migrants, as well as net lifetime migrants.

Add minor changes to the script.